Create Data dig Project - Marvel database
I had a marvel characters database, which included name, popularity, height in meters, weight in kg, and hometown. I worked out the characters' BMI with the given heights and weights, to see who's healthy, overweight and underweight. I also used the popularity ranking to figure out which characters need a PR team, and also used the grouping of the hometowns for the poorly ranking characters, to assess how we could split the PR team to go to the hometowns.